backend/pyproject.toml
[project]
name = "vkr-english-platform"
version = "0.1.0"
description = "Modular monolith backend for RU-native users learning English"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "sqlalchemy>=2.0.32",
  "psycopg[binary]>=3.2.0",
  "alembic>=1.13.2",
  "email-validator>=2.2.0",
  "PyJWT>=2.9.0",
  "pytest>=8.3.0",
  "httpx>=0.27.0",
  "celery[redis]>=5.4.0",
  "redis>=5.0.0",
  "flower>=2.0.0"
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6.0",
  "mypy>=1.11.0",
  "pre-commit>=3.8.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP", "B", "A", "SIM", "TCH"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
